+++ src/stactools/core/__init__.py
@@ -5,5 +5,19 @@ from stactools.core.copy import (move_asset_file_to_item, move_assets,
                                  move_all_assets, copy_catalog)
 from stactools.core.layout import layout_catalog
 from stactools.core.merge import (merge_items, merge_all_items)
+from stactools.core.add import add_item
+from stactools.core.addraster import add_raster_to_item
 
-__version__ = "0.2.1"
+__all__ = [
+    "add_item",
+    "add_raster_to_item",
+    "copy_catalog",
+    "layout_catalog",
+    "merge_all_items",
+    "merge_items",
+    "move_asset_file_to_item",
+    "move_assets",
+    "move_all_assets",
+    "use_fsspec",
+]
+__version__ = "0.2.6"