+++ src/stactools/core/io/xml.py
@@ -11,12 +11,13 @@ class XmlElement:
     """Thin wrapper around lxml etree.Element with some
     convenience functions
     """
+
     def __init__(self, element: lxmlElement):
         self.element = element
 
     @lru_cache(maxsize=100)
     def find(self, xpath: str) -> Optional["XmlElement"]:
-        node = self.element.find(xpath, self.element.nsmap)
+        node = self.element.find(xpath, self.element.nsmap)  # type: ignore
         return None if node is None else XmlElement(node)
 
     def find_or_throw(
@@ -30,8 +31,8 @@ class XmlElement:
     @lru_cache(maxsize=100)
     def findall(self, xpath: str) -> List["XmlElement"]:
         return [
-            XmlElement(e)
-            for e in self.element.findall(xpath, self.element.nsmap)
+            XmlElement(e) for e in self.element.findall(
+                xpath, self.element.nsmap)  # type: ignore
         ]
 
     @lru_cache(maxsize=100)
@@ -52,8 +53,14 @@ class XmlElement:
         return None if node is None else node.get_attr(attr)
 
     @property
-    def text(self):
-        return self.element.text
+    def text(self) -> Optional[str]:
+        if isinstance(self.element.text, str):
+            return self.element.text
+        elif isinstance(self.element.text, bytes):
+            return str(self.element.text, encoding='utf-8')
+        else:
+            assert self.element.text is None
+            return None
 
     @lru_cache(maxsize=100)
     def get_attr(self, attr: str) -> Optional[str]: