+++ tests/core/test_io.py
@@ -7,6 +7,7 @@ from stactools.core import use_fsspec
 
 
 class IOTest(unittest.TestCase):
+
     def test_fsspec_io(self):
         use_fsspec()